requirements.txt
matplotlib
numpy
pylint
tensorboard
torchaudio
torchvision
